Scripts/requirements.txt
requests~=2.31.0
discord.py~=2.3.2
python-dotenv~=1.0.1